Where I am
This is the point I started using GitHub for this project, taking it more seriously.
In this project there are currently 3 files, message, client and server
The message class is there for ease of use to easily input data, whether int, char, string or any other variable, support for lists is coming soon. The client can handle specific messages differently, can connect to a server and send the messages easily. The same goes for the server, this is mainly it. The server also can handle multiple clients and then there are functions like send all clients and so on. This is trying to be a good project and will result in a final good library.