LNet & ENet
LNetStream is the project which we use ENet for communication. The code is relatively similar but there are things to change like how the type is now 2 bytes, but there is another type called channel which is almost one byte (until 254) so we have 16646144 message types, which is enough for now.
The only problem is to make it nice for users to choose where and how to send the message...